+++ packages/plugins/Pins/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## unpublished
 
+- Deprecated: Using `movable` with a `boolean` has been deprecated. Please use the parameter with a string. For more information, see the documentation of the package.
 - Feature: Add new configuration parameter `initial` to be able to add a pin initially to the map.
-- Feature: If `movable` is set to `false`, the style of the cursor is set to `not-allowed` when hovering the pin.
+- Feature: If `movable` is set to `none` or `false`, the style of the cursor is set to `not-allowed` when hovering the pin.
 
 ## 1.0.0